refactor: move task classes to a folder
