+++ TimeInABottle.Core/Contracts/Services/IDaoService.cs
@@ -4,7 +4,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 using TimeInABottle.Core.Helpers;
-using TimeInABottle.Core.Models;
+using TimeInABottle.Core.Models.Tasks;
 
 namespace TimeInABottle.Core.Contracts.Services;
 /// <summary>
