+++ TimeInABottle.Core/Contracts/Services/IWeatherService.cs
@@ -1,0 +1,10 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace TimeInABottle.Core.Contracts.Services;
+internal interface IWeatherService
+{
+}
