+++ TimeInABottle.Core/Helpers/TaskListSorter.cs
@@ -3,7 +3,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
-using TimeInABottle.Core.Models;
+using TimeInABottle.Core.Models.Tasks;
 
 namespace TimeInABottle.Core.Helpers;
 /// <summary>
