+++ TimeInABottle.Core/Models/Filters/CompositeFilter.cs
@@ -6,6 +6,7 @@ using System.Linq;
 using System.Reflection.Metadata.Ecma335;
 using System.Text;
 using System.Threading.Tasks;
+using TimeInABottle.Core.Models.Tasks;
 
 
 namespace TimeInABottle.Core.Models.Filters;
