+++ TimeInABottle.Core/Models/Filters/TimeFilter.cs
@@ -4,6 +4,7 @@ using System.ComponentModel;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
+using TimeInABottle.Core.Models.Tasks;
 
 namespace TimeInABottle.Core.Models.Filters;
 /// <summary>
