+++ TimeInABottle.Core/Models/Tasks/DerivedTask.cs
@@ -5,7 +5,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace TimeInABottle.Core.Models;
+namespace TimeInABottle.Core.Models.Tasks;
 /// <summary>
 /// Represents a derived task with specific properties and methods.
 /// </summary>
