+++ TimeInABottle.Core/secret.config
@@ -1,0 +1,4 @@
+﻿<?xml version="1.0" encoding="utf-8" ?>
+<configuration>
+	<APIkey>8YK9AGLfawAKwfBZs0v8aQD4qKNCMpuZ</APIkey>
+</configuration>
