+++ TimeInABottle/Views/DashboardPage.xaml
@@ -5,7 +5,7 @@
     xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
     xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
     xmlns:helpers="using:TimeInABottle.Helpers"
-    xmlns:models="using:TimeInABottle.Core.Models"
+    xmlns:task="using:TimeInABottle.Core.Models.Tasks"
     xmlns:xaml="using:Microsoft.UI.Xaml"
     mc:Ignorable="d">
 
@@ -14,7 +14,7 @@
         <helpers:TimeToStringConverter x:Key="TimeToStringConverter"/>
         <helpers:TaskToNotificationConverter x:Key="TaskToNotificationConverter"/>
 
-        <DataTemplate x:Key ="ListItem" x:DataType="models:ITask">
+        <DataTemplate x:Key ="ListItem" x:DataType="task:ITask">
             <StackPanel Orientation="Horizontal" Style="{StaticResource BaseListItemContainerStyle}">
                 <TextBlock Style="{StaticResource ListSubTitleStyle}">
                     <Run Text="{x:Bind Start, Converter={StaticResource TimeToStringConverter}}"/>: <Run Text="{x:Bind Name}"/>
