+++ TimeInABottle/Views/TaskListDetailControl.xaml.cs
@@ -1,7 +1,6 @@
 ﻿using Microsoft.UI.Xaml;
 using Microsoft.UI.Xaml.Controls;
-
-using TimeInABottle.Core.Models;
+using TimeInABottle.Core.Models.Tasks;
 
 namespace TimeInABottle.Views;
